Pull tracetable from GitHub and not crds
